Split ScannerEx into Highway and NumKong backends
