+++ src/core/calc/scanner_nk.h
@@ -1,10 +1,10 @@
-// Declares the calc-engine-backed top-k scanner API.
+// Declares the NumKong-backed scanner implementation.
 
 #pragma once
-#include "core/calc/calc_engine.h"
+
 #include "core/calc/dist_item.h"
 #include "core/utils/bitset_filter.h"
-#include "utils/shared_types.h"
+
 #include <cstdint>
 #include <vector>
 
@@ -12,23 +12,13 @@ namespace sketch2 {
 
 class DatasetReader;
 
-const char* calc_engine_name(CalcEngine engine);
-
-// ScannerEx provides the same top-k search as Scanner but delegates metric
-// score computation to the calc engine layer instead of the
-// Singleton-dispatched hand-written SIMD kernels.
-class ScannerEx {
+class ScannerNk {
 public:
-    explicit ScannerEx(CalcEngine engine = CalcEngine::highway);
-
     Ret find(const DatasetReader& dataset, size_t count, const uint8_t* vec,
         std::vector<uint64_t>& result) const;
 
     Ret find_items(const DatasetReader& dataset, size_t count, const uint8_t* vec,
         std::vector<DistItem>& result, const BitsetFilter* bitset = nullptr) const;
-
-private:
-    CalcEngine engine_;
 };
 
 } // namespace sketch2
